+++ lib/rubocop/cop/design_cops.rb
@@ -1,8 +1,0 @@
-require_relative "design/argument_counting"
-require_relative "design/keyword_arguments"
-require_relative "design/positional_arguments"
-require_relative "design/total_arguments"
-require_relative "design/agent_noun_class_name"
-require_relative "design/nested_method_calls"
-require_relative "design/spec_comment"
-require_relative "design/file_utils_inclusion"